requirements.txt
Flask==3.0.0
python-docx==1.1.0
docxcompose==1.4.0
Werkzeug==3.0.1
Pillow==10.1.0
pdf2image==1.16.3
Flask-SocketIO==5.3.5
python-socketio==5.10.0
Flask-Compress==1.14
